news-agent/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-community
langchain-google-genai
tavily-python
python-dotenv
fastapi              # optional — if you want API later
uvicorn              # optional
# Terminal Formatting
rich>=14.3.1